README.md
The README.md file for this project. There was a little delay for this commit, as I accidentally pressed the pause/break key on my keyboard, and couldn't figure out why everything was acting so wonky. I eventually fixed it, but I wonder what the purpose of this key even is. This file is in Markdown (*.md) format. File_version=1;